+++ test/clj/metosin/postgres/types_test.clj
@@ -5,12 +5,6 @@
 
 
 (deftest pgobject->clj-test
-  (testing "keyword"
-    (is (= :foo/bar (pgobject->clj (->PGobject keyword-type-name "(foo/bar)"))))
-    (is (= :foo/bar (pgobject->clj (jdbc/sql-value :foo/bar))))
-    (is (= :bar (pgobject->clj (->PGobject keyword-type-name "(bar)"))))
-    (is (= :bar (pgobject->clj (jdbc/sql-value :bar)))))
-
   (testing "json"
     (is (= {:foo "a"} (pgobject->clj (->PGobject "json" "{\"foo\":\"a\"}"))))
     (is (= {:foo "a"} (pgobject->clj (jdbc/sql-value {:foo "a"}))))